% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-portal-rodents.R
\name{extract_portal_year}
\alias{extract_portal_year}
\title{Extract observations for a given year}
\usage{
extract_portal_year(granivores, year = 1985)
}
\arguments{
\item{granivores}{Dataframe of year, species, and weight}

\item{year}{Which year}
}
\value{
dataframe of species ID and weight (1 for just abundances, power for power) for all records from that year.
}
\description{
Extract observations for a given year
}
