% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-mete-sims.R
\name{generate_sims}
\alias{generate_sims}
\title{Generate simulated METE communities}
\usage{
generate_sims(mete_distribution, nsims = 10000, return_sims = F,
  store = T, store_path)
}
\arguments{
\item{mete_distribution}{a meteDist object}

\item{nsims}{how many sims to pull}

\item{return_sims}{logical: return sims or just evaluate silently?}

\item{store}{logical: store or not?}

\item{store_path}{where to store the sims (end in .Rds)}
}
\value{
list of 3: [\link{1}] is the type (sad, ipd, sipd), [\link{2}] is empirical vector; [\link{3}] is a list of nsims simulated vectors
}
\description{
Generate simulations (draws from METE probability distributions) for a community.
}
