% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapping-functions.R
\name{get_community_prob}
\alias{get_community_prob}
\title{Get probabilities of empirical coming from mete for a whole community.}
\usage{
get_community_prob(community_data_frame, sim_storage_path)
}
\arguments{
\item{community_data_frame}{dataframe of species, abundance, power for all individuals}

\item{sim_storage_path}{storage path for stashing sims}
}
\value{
list of p of SAD, p of IPD coming from METE predictions based on community data
}
\description{
Currently uses loglikelihood and assumes the loglikelihoods of sim distributions|METE are normally distributed.
}
