% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-test-stats.R
\name{get_ksD}
\alias{get_ksD}
\title{Get a Kolmogorov-Smirnov D
Get a Kolmogorov-Smirnov D to compare an observed distribution to a predicted METE distribution.}
\usage{
get_ksD(obs_ranked, meteR_dist)
}
\arguments{
\item{obs_ranked}{Vector of observed values, ranked decreasing}

\item{meteR_dist}{The meteR meteDist object for comparison}
}
\value{
ks-D value
}
\description{
Get a Kolmogorov-Smirnov D
Get a Kolmogorov-Smirnov D to compare an observed distribution to a predicted METE distribution.
}
