% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-test-stats.R
\name{get_logLik}
\alias{get_logLik}
\title{Get a logLik}
\usage{
get_logLik(obs_ranked, meteR_dist)
}
\arguments{
\item{obs_ranked}{Vector of observed values, ranked decreasing}

\item{meteR_dist}{The meteR meteDist object for comparison}
}
\value{
logLik value
}
\description{
Get the log-likelihood of an observed vector (either empirical or simulated from METE) and a meteR meteDist object
}
\details{
Code adapted from the logLik function in meteR.
}
