% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-rank-quantiles.R
\name{get_quantile}
\alias{get_quantile}
\title{Get rank quantile for a single rank in a distribution}
\usage{
get_quantile(all_values, empirical_row)
}
\arguments{
\item{all_values}{vector of all values, including the empirical value}

\item{empirical_row}{index of the empirical value}
}
\value{
probability of empirical value given ecdf of sim values
}
\description{
Get the probability of the empirical value compared to a
vector of sim values.
}
\details{
A helper function for `get_rank_quantile``
}
