% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-rank-quantiles.R
\name{get_rank_quantile}
\alias{get_rank_quantile}
\title{Get rank quantiles for a whole distribution}
\usage{
get_rank_quantile(sim_dists)
}
\arguments{
\item{sim_dists}{from \code{generate_sims}}
}
\value{
Vector of quantiles of observed vs. sim distribution
}
\description{
Get the probability of the empirical value at all ranks in a
distribution relative to the distribution of values for each rank
for sims drawn from the METE prediction
}
