% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate-rank-quantiles.R
\name{get_theoretical_quantile}
\alias{get_theoretical_quantile}
\title{Get quantile of empirical value compared to theoretical sim distribution}
\usage{
get_theoretical_quantile(all_values, empirical_row,
  theoretical_dist = "norm")
}
\arguments{
\item{all_values}{vector of all values, including the empirical value}

\item{empirical_row}{index of the empirical value}

\item{theoretical_dist}{"norm" to fit a normal distribution to the sim values}
}
\value{
probability of empirical value given probability function generated from sim values
}
\description{
Get quantile of empirical value compared to theoretical sim distribution
}
