% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-mete-predictions.R
\name{make_iIPDs}
\alias{make_iIPDs}
\title{Generate the iIPDs for a community as a list}
\usage{
make_iIPDs(meteESF)
}
\arguments{
\item{meteESF}{the meteESF object for the community}
}
\value{
a list of meteDists for the iIPDs (also called SIPDs) for all species in a community.
}
\description{
Using meteR, generate the iIPDs for all species in a community
}
