% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-mete-predictions.R
\name{make_mete_distribution}
\alias{make_mete_distribution}
\title{Generate METE predictions for SAD, IPD, iIPD}
\usage{
make_mete_distribution(community_data, distribution = "SAD")
}
\arguments{
\item{community_data}{dataframe of species ID and weights (1 for abundance only) by individuals}

\item{distribution}{"SAD", "IPD", or "iIPD"; which METE distribution to make the prediction for}
}
\value{
appropriate meteDist object (see meteR documentation for details)
}
\description{
Using meteR, generate the predicted distribution (SAD, IPD, iIPD) for a given community dataset
}
