% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-mete-predictions.R
\name{make_mete_distributions}
\alias{make_mete_distributions}
\title{Generate the METE distributions for a community as a list}
\usage{
make_mete_distributions(this_ESF, power_dists = T)
}
\arguments{
\item{power_dists}{logical: whether or not to include IPD and SIPD. Use F if no power data.}

\item{meteESF}{the meteESF object for the community}
}
\value{
a list of meteDists for the community: sad, ipd, sipds
}
\description{
Using meteR, generate the iIPDs for all species in a community
}
