% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-mete-predictions.R
\name{make_mete_ESF}
\alias{make_mete_ESF}
\title{Generate METE predictions for SAD, IPD, iIPD}
\usage{
make_mete_ESF(community_data)
}
\arguments{
\item{community_data}{dataframe of species ID and weights (1 for abundance only) by individuals}
}
\value{
meteESF
}
\description{
Using meteR, generate the ESF for a given community dataset
}
