% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-mete-sims.R
\name{pull_sim}
\alias{pull_sim}
\title{Sample a mete distribution}
\usage{
pull_sim(mete_distribution, svar_n, svar_sum)
}
\arguments{
\item{mete_distribution}{a meteDist object}

\item{svar_n}{how long the vector should be}

\item{svar_sum}{what the vector should sum to}
}
\value{
a vector drawn from mete_distribution, of length svar_n and that sums to svar_sum.
}
\description{
Draws a vector from a mete probability distribution that conforms to given state variables.
Called by generate_sims().
}
