% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-portal-rodents.R
\name{load_rodent_data}
\alias{load_rodent_data}
\title{Load Portal granivores data}
\usage{
load_rodent_data(treatment = "control", compute_power = TRUE)
}
\arguments{
\item{treatment}{Whether to load "exclosure" or "control" rodents}

\item{compute_power}{Whether to compute power. Note that if TRUE, individuals missing weight measurements will be removed, and they will not be removed if FALSE.}
}
\value{
dataframe of year, species ID, abundances (all 1, for meteR), and power (NA if report_power = F, metabolic rate otherwise) for all individuals captured.
}
\description{
All individual captures and the year. Reported as individuals and optionally power (power column is estimated metabolic rate, as 5.69 * (biomass ^ 0.75), after White et al 2004).
Non-granivorous species, and individuals not identified to species, are removed.
}
